+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/impl/CertifiedProductSearchResultDAOImpl.java
@@ -226,6 +226,9 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
 		}
 		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
+		}
 		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
 			queryStr += " AND (visible_on_chpl = true ) ";
@@ -294,6 +297,11 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
+		
 		return query;
 	}
 	
@@ -414,6 +422,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
 		}
 		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
+		}
+		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
 			queryStr += " AND (visible_on_chpl = true ) ";
 		} else if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("no")){
@@ -481,6 +493,11 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
+		
 		return query;
 		
 	}
@@ -609,6 +626,9 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
 		}
 		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
+		}
 		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
 			queryStr += " AND (visible_on_chpl = true ) ";
@@ -680,6 +700,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
 		return query;
 	}
 	
@@ -717,6 +741,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		
 		if (searchRequest.getVersion() != null) {
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version))";
+		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
 		}
 		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
@@ -780,6 +808,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
+		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
 		}
 		
 		return query;
@@ -872,6 +904,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
 		}
 		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
+		}
+		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
 			queryStr += " AND (visible_on_chpl = true ) ";
 		} else if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("no")){
@@ -930,6 +966,11 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
+		
 		return query;
 	}
 	
@@ -984,6 +1025,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		
 		if (searchRequest.getVersion() != null) {
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
+		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
 		}
 		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
@@ -1044,6 +1089,11 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
+		
 		return query;
 		
 	}
@@ -1107,6 +1157,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
 		}
 		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
+		}
+		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
 			queryStr += " AND (visible_on_chpl = true ) ";
 		} else if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("no")){
@@ -1168,6 +1222,11 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
 		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
+		}
+		
 		return query;
 	}
 	
@@ -1205,6 +1264,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		
 		if (searchRequest.getVersion() != null) {
 			queryStr += " AND (UPPER(product_version) LIKE UPPER(:version)) ";
+		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			queryStr += " AND (UPPER(certification_status_name) LIKE UPPER(:certificationStatus))";
 		}
 		
 		if (searchRequest.getVisibleOnCHPL().toLowerCase().startsWith("yes") ){
@@ -1260,6 +1323,10 @@ public class CertifiedProductSearchResultDAOImpl extends BaseDAOImpl implements
 		
 		if (searchRequest.getVersion() != null) {
 			query.setParameter("version", "%"+searchRequest.getVersion()+"%");
+		}
+		
+		if(searchRequest.getCertificationStatus() != null) {
+			query.setParameter("certificationStatus", searchRequest.getCertificationStatus());
 		}
 		return query;
 	}